+++ variants/models.py
@@ -188,6 +188,9 @@ class CaseAwareProject(Project):
     def casealignmentstats(self):
         return [case.latest_variant_set().casealignmentstats for case in self.case_set.all()]
 
+    def get_annotation_count(self):
+        return sum(case.get_annotation_count() for case in self.case_set.all())
+
 
 class SmallVariant(models.Model):
     """"Information of a single variant, knows its case."""